+++ src/libs/notion/blocks/convertResponseToBlock/childDatabase.ts
@@ -1,6 +1,6 @@
 import { isFullDatabase } from "@notionhq/client";
 
-import { retrieveDatabase } from "../../databases.js";
+import { retrieveDatabase } from "../../database/retrieve.js";
 
 import type { ChildDatabaseBlockObject } from "../../../../types/notion/blocks/childDatabase.js";
 import type { ChildDatabaseBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints.js";
