+++ src/libs/notion/index.ts
@@ -38,7 +38,7 @@ export * from "./richText/mention.js";
 export * from "./richText/richText.js";
 export * from "./richText/richTextItem.js";
 export * from "./richText/text.js";
-export * from "./databases.js";
+export * from "./database/retrieve.js";
 export * from "./pages/pages.js";
 export * from "./parent.js";
 export * from "./comments.js";
